Move mocks back inside listpage tests
